Migrate size_calc FieldType methods into subpackage
